Helicopter now moves and rotates as expected... kinda

+++ HornetsHfd-A2/common/src/main/java/org/csc133/a2/interfaces/Steerable.java
@@ -2,8 +2,8 @@ package org.csc133.a2.interfaces;
 
 public interface Steerable {
 
-    void turnLeft(double amount);
+    void turnLeft(float amount);
 
-    void turnRight(double amount);
+    void turnRight(float amount);
 
 }
